+++ src/com/ipm/web/interfaces/dao/SkillDao.java
@@ -6,7 +6,7 @@ import com.ipm.web.dto.Skill;
 
 public interface SkillDao {
 
-	public List<Skill> getSkills(String username, int projectId);
+	public List<Skill> getSkills(String username, long projectId);
 
 	public void createSkill(Skill skill);
